+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/java/com/liferay/dynamic/data/lists/form/web/internal/portlet/action/AddRecordSetMVCActionCommand.java
@@ -15,51 +15,8 @@
 package com.liferay.dynamic.data.lists.form.web.internal.portlet.action;
 
 import com.liferay.dynamic.data.lists.form.web.constants.DDLFormPortletKeys;
-import com.liferay.dynamic.data.lists.form.web.internal.converter.DDLFormRulesToDDMFormRulesConverter;
-import com.liferay.dynamic.data.lists.form.web.internal.converter.model.DDLFormRule;
-import com.liferay.dynamic.data.lists.model.DDLRecordSet;
-import com.liferay.dynamic.data.lists.model.DDLRecordSetConstants;
-import com.liferay.dynamic.data.lists.model.DDLRecordSetSettings;
-import com.liferay.dynamic.data.lists.service.DDLRecordSetService;
-import com.liferay.dynamic.data.mapping.exception.StructureDefinitionException;
-import com.liferay.dynamic.data.mapping.exception.StructureLayoutException;
-import com.liferay.dynamic.data.mapping.form.values.query.DDMFormValuesQuery;
-import com.liferay.dynamic.data.mapping.form.values.query.DDMFormValuesQueryFactory;
-import com.liferay.dynamic.data.mapping.io.DDMFormJSONDeserializer;
-import com.liferay.dynamic.data.mapping.io.DDMFormLayoutJSONDeserializer;
-import com.liferay.dynamic.data.mapping.io.DDMFormValuesJSONDeserializer;
-import com.liferay.dynamic.data.mapping.model.DDMForm;
-import com.liferay.dynamic.data.mapping.model.DDMFormLayout;
-import com.liferay.dynamic.data.mapping.model.DDMFormRule;
-import com.liferay.dynamic.data.mapping.model.DDMStructure;
-import com.liferay.dynamic.data.mapping.model.DDMStructureConstants;
-import com.liferay.dynamic.data.mapping.model.Value;
-import com.liferay.dynamic.data.mapping.service.DDMStructureService;
-import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
-import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
-import com.liferay.dynamic.data.mapping.storage.StorageType;
-import com.liferay.dynamic.data.mapping.util.DDMFormFactory;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.json.JSONDeserializer;
-import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseTransactionalMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextFactory;
-import com.liferay.portal.kernel.service.WorkflowDefinitionLinkLocalService;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.PortalUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Locale;
-import java.util.Map;
-import java.util.Objects;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
@@ -81,291 +38,17 @@ import org.osgi.service.component.annotations.Reference;
 public class AddRecordSetMVCActionCommand
 	extends BaseTransactionalMVCActionCommand {
 
-	protected DDMStructure addDDMStructure(
-			ActionRequest actionRequest, DDMFormValues settingsDDMFormValues)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		long groupId = ParamUtil.getLong(actionRequest, "groupId");
-		String structureKey = ParamUtil.getString(
-			actionRequest, "structureKey");
-		String storageType = getStorageType(settingsDDMFormValues);
-		String name = ParamUtil.getString(actionRequest, "name");
-		String description = ParamUtil.getString(actionRequest, "description");
-		DDMForm ddmForm = getDDMForm(actionRequest);
-		DDMFormLayout ddmFormLayout = getDDMFormLayout(actionRequest);
-
-		ServiceContext serviceContext = ServiceContextFactory.getInstance(
-			DDMStructure.class.getName(), actionRequest);
-
-		return ddmStructureService.addStructure(
-			groupId, DDMStructureConstants.DEFAULT_PARENT_STRUCTURE_ID,
-			PortalUtil.getClassNameId(DDLRecordSet.class), structureKey,
-			getLocalizedMap(themeDisplay.getSiteDefaultLocale(), name),
-			getLocalizedMap(themeDisplay.getSiteDefaultLocale(), description),
-			ddmForm, ddmFormLayout, storageType,
-			DDMStructureConstants.TYPE_AUTO, serviceContext);
-	}
-
-	protected DDLRecordSet addRecordSet(
-			ActionRequest actionRequest, long ddmStructureId)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		long groupId = ParamUtil.getLong(actionRequest, "groupId");
-		String recordSetKey = ParamUtil.getString(
-			actionRequest, "recordSetKey");
-		String name = ParamUtil.getString(actionRequest, "name");
-		String description = ParamUtil.getString(actionRequest, "description");
-
-		ServiceContext serviceContext = ServiceContextFactory.getInstance(
-			DDLRecordSet.class.getName(), actionRequest);
-
-		return ddlRecordSetService.addRecordSet(
-			groupId, ddmStructureId, recordSetKey,
-			getLocalizedMap(themeDisplay.getSiteDefaultLocale(), name),
-			getLocalizedMap(themeDisplay.getSiteDefaultLocale(), description),
-			DDLRecordSetConstants.MIN_DISPLAY_ROWS_DEFAULT,
-			DDLRecordSetConstants.SCOPE_FORMS, serviceContext);
-	}
-
 	@Override
 	protected void doTransactionalCommand(
 			ActionRequest actionRequest, ActionResponse actionResponse)
 		throws Exception {
 
-		DDMFormValues settingsDDMFormValues = getSettingsDDMFormValues(
-			actionRequest);
-
-		DDMStructure ddmStructure = addDDMStructure(
-			actionRequest, settingsDDMFormValues);
-
-		DDLRecordSet recordSet = addRecordSet(
-			actionRequest, ddmStructure.getStructureId());
-
-		updateRecordSetSettings(
-			actionRequest, recordSet, settingsDDMFormValues);
+		addUpdateRecordSetMVCCommandHelper.addRecordSet(
+			actionRequest, actionResponse);
 	}
 
-	protected DDMForm getDDMForm(ActionRequest actionRequest)
-		throws PortalException {
-
-		try {
-			String definition = ParamUtil.getString(
-				actionRequest, "definition");
-
-			DDMForm ddmForm = ddmFormJSONDeserializer.deserialize(definition);
-
-			List<DDMFormRule> ddmFormRules = getDDMFormRules(actionRequest);
-
-			ddmForm.setDDMFormRules(ddmFormRules);
-
-			return ddmForm;
-		}
-		catch (PortalException pe) {
-			throw new StructureDefinitionException(pe);
-		}
-	}
-
-	protected DDMFormLayout getDDMFormLayout(ActionRequest actionRequest)
-		throws PortalException {
-
-		try {
-			String layout = ParamUtil.getString(actionRequest, "layout");
-
-			return ddmFormLayoutJSONDeserializer.deserialize(layout);
-		}
-		catch (PortalException pe) {
-			throw new StructureLayoutException(pe);
-		}
-	}
-
-	protected List<DDMFormRule> getDDMFormRules(ActionRequest actionRequest)
-		throws PortalException {
-
-		String rules = ParamUtil.getString(actionRequest, "rules");
-
-		if (Validator.isNull(rules) || Objects.equals("[]", rules)) {
-			return Collections.emptyList();
-		}
-
-		JSONDeserializer<DDLFormRule[]> jsonDeserializer =
-			jsonFactory.createJSONDeserializer();
-
-		DDLFormRule[] ddlFormRules = jsonDeserializer.deserialize(
-			rules, DDLFormRule[].class);
-
-		return ddlFormRulesToDDMFormRulesConverter.convert(
-			ListUtil.toList(ddlFormRules));
-	}
-
-	protected Map<Locale, String> getLocalizedMap(Locale locale, String value) {
-		Map<Locale, String> localizedMap = new HashMap<>();
-
-		localizedMap.put(locale, value);
-
-		return localizedMap;
-	}
-
-	protected DDMFormValues getSettingsDDMFormValues(
-			ActionRequest actionRequest)
-		throws PortalException {
-
-		String serializedSettingsDDMFormValues = ParamUtil.getString(
-			actionRequest, "serializedSettingsDDMFormValues");
-
-		DDMForm ddmForm = DDMFormFactory.create(DDLRecordSetSettings.class);
-
-		DDMFormValues settingsDDMFormValues =
-			ddmFormValuesJSONDeserializer.deserialize(
-				ddmForm, serializedSettingsDDMFormValues);
-
-		return settingsDDMFormValues;
-	}
-
-	protected String getStorageType(DDMFormValues ddmFormValues)
-		throws PortalException {
-
-		DDMFormValuesQuery ddmFormValuesQuery =
-			ddmFormValuesQueryFactory.create(ddmFormValues, "/storageType");
-
-		DDMFormFieldValue ddmFormFieldValue =
-			ddmFormValuesQuery.selectSingleDDMFormFieldValue();
-
-		Value value = ddmFormFieldValue.getValue();
-
-		String storageType = value.getString(ddmFormValues.getDefaultLocale());
-
-		if (Validator.isNull(storageType)) {
-			storageType = StorageType.JSON.toString();
-		}
-
-		return storageType;
-	}
-
-	protected String getWorkflowDefinition(DDMFormValues ddmFormValues)
-		throws PortalException {
-
-		DDMFormValuesQuery ddmFormValuesQuery =
-			ddmFormValuesQueryFactory.create(
-				ddmFormValues, "/workflowDefinition");
-
-		DDMFormFieldValue ddmFormFieldValue =
-			ddmFormValuesQuery.selectSingleDDMFormFieldValue();
-
-		Value value = ddmFormFieldValue.getValue();
-
-		return value.getString(ddmFormValues.getDefaultLocale());
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDLFormRulesToDDMFormRulesConverter(
-		DDLFormRulesToDDMFormRulesConverter
-			ddlFormRulesToDDMFormRulesConverter) {
-
-		this.ddlFormRulesToDDMFormRulesConverter =
-			ddlFormRulesToDDMFormRulesConverter;
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDLRecordSetService(
-		DDLRecordSetService ddlRecordSetService) {
-
-		this.ddlRecordSetService = ddlRecordSetService;
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDMFormJSONDeserializer(
-		DDMFormJSONDeserializer ddmFormJSONDeserializer) {
-
-		this.ddmFormJSONDeserializer = ddmFormJSONDeserializer;
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDMFormLayoutJSONDeserializer(
-		DDMFormLayoutJSONDeserializer ddmFormLayoutJSONDeserializer) {
-
-		this.ddmFormLayoutJSONDeserializer = ddmFormLayoutJSONDeserializer;
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDMFormValuesJSONDeserializer(
-		DDMFormValuesJSONDeserializer ddmFormValuesJSONDeserializer) {
-
-		this.ddmFormValuesJSONDeserializer = ddmFormValuesJSONDeserializer;
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDMFormValuesQueryFactory(
-		DDMFormValuesQueryFactory ddmFormValuesQueryFactory) {
-
-		this.ddmFormValuesQueryFactory = ddmFormValuesQueryFactory;
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDMStructureService(
-		DDMStructureService ddmStructureService) {
-
-		this.ddmStructureService = ddmStructureService;
-	}
-
-	@Reference(unbind = "-")
-	protected void setJSONFactory(JSONFactory jsonFactory) {
-		this.jsonFactory = jsonFactory;
-	}
-
-	@Reference(unbind = "-")
-	protected void setWorkflowDefinitionLinkLocalService(
-		WorkflowDefinitionLinkLocalService workflowDefinitionLinkLocalService) {
-
-		this.workflowDefinitionLinkLocalService =
-			workflowDefinitionLinkLocalService;
-	}
-
-	protected void updateRecordSetSettings(
-			ActionRequest actionRequest, DDLRecordSet recordSet,
-			DDMFormValues settingsDDMFormValues)
-		throws PortalException {
-
-		ddlRecordSetService.updateRecordSet(
-			recordSet.getRecordSetId(), settingsDDMFormValues);
-
-		updateWorkflowDefinitionLink(
-			actionRequest, recordSet, settingsDDMFormValues);
-	}
-
-	protected void updateWorkflowDefinitionLink(
-			ActionRequest actionRequest, DDLRecordSet recordSet,
-			DDMFormValues ddmFormValues)
-		throws PortalException {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		long groupId = ParamUtil.getLong(actionRequest, "groupId");
-
-		String workflowDefinition = getWorkflowDefinition(ddmFormValues);
-
-		workflowDefinitionLinkLocalService.updateWorkflowDefinitionLink(
-			themeDisplay.getUserId(), themeDisplay.getCompanyId(), groupId,
-			DDLRecordSet.class.getName(), recordSet.getRecordSetId(), 0,
-			workflowDefinition);
-	}
-
-	protected DDLFormRulesToDDMFormRulesConverter
-		ddlFormRulesToDDMFormRulesConverter;
-	protected DDLRecordSetService ddlRecordSetService;
-	protected DDMFormJSONDeserializer ddmFormJSONDeserializer;
-	protected DDMFormLayoutJSONDeserializer ddmFormLayoutJSONDeserializer;
-	protected DDMFormValuesJSONDeserializer ddmFormValuesJSONDeserializer;
-	protected DDMFormValuesQueryFactory ddmFormValuesQueryFactory;
-	protected DDMStructureService ddmStructureService;
-	protected JSONFactory jsonFactory;
-	protected volatile WorkflowDefinitionLinkLocalService
-		workflowDefinitionLinkLocalService;
+	@Reference
+	protected AddUpdateRecordSetMVCCommandHelper
+		addUpdateRecordSetMVCCommandHelper;
 
 }
