+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/java/com/liferay/dynamic/data/lists/form/web/internal/portlet/action/UpdateRecordSetMVCActionCommand.java
@@ -15,24 +15,14 @@
 package com.liferay.dynamic.data.lists.form.web.internal.portlet.action;
 
 import com.liferay.dynamic.data.lists.form.web.constants.DDLFormPortletKeys;
-import com.liferay.dynamic.data.lists.model.DDLRecordSet;
-import com.liferay.dynamic.data.lists.model.DDLRecordSetConstants;
-import com.liferay.dynamic.data.mapping.model.DDMForm;
-import com.liferay.dynamic.data.mapping.model.DDMFormLayout;
-import com.liferay.dynamic.data.mapping.model.DDMStructure;
-import com.liferay.dynamic.data.mapping.model.DDMStructureConstants;
-import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
+import com.liferay.portal.kernel.portlet.bridges.mvc.BaseTransactionalMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextFactory;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.WebKeys;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Bruno Basto
@@ -46,68 +36,19 @@ import org.osgi.service.component.annotations.Component;
 	service = MVCActionCommand.class
 )
 public class UpdateRecordSetMVCActionCommand
-	extends AddRecordSetMVCActionCommand {
+	extends BaseTransactionalMVCActionCommand {
 
 	@Override
 	protected void doTransactionalCommand(
 			ActionRequest actionRequest, ActionResponse actionResponse)
 		throws Exception {
 
-		DDMStructure ddmStructure = updateDDMStructure(actionRequest);
-
-		DDLRecordSet recordSet = updateRecordSet(
-			actionRequest, ddmStructure.getStructureId());
-
-		DDMFormValues settingsDDMFormValues = getSettingsDDMFormValues(
-			actionRequest);
-
-		updateRecordSetSettings(
-			actionRequest, recordSet, settingsDDMFormValues);
+		addUpdateRecordSetMVCCommandHelper.updateRecordSet(
+			actionRequest, actionResponse);
 	}
 
-	protected DDMStructure updateDDMStructure(ActionRequest actionRequest)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		long ddmStructureId = ParamUtil.getLong(
-			actionRequest, "ddmStructureId");
-		String name = ParamUtil.getString(actionRequest, "name");
-		String description = ParamUtil.getString(actionRequest, "description");
-		DDMForm ddmForm = getDDMForm(actionRequest);
-		DDMFormLayout ddmFormLayout = getDDMFormLayout(actionRequest);
-
-		ServiceContext serviceContext = ServiceContextFactory.getInstance(
-			DDMStructure.class.getName(), actionRequest);
-
-		return ddmStructureService.updateStructure(
-			ddmStructureId, DDMStructureConstants.DEFAULT_PARENT_STRUCTURE_ID,
-			getLocalizedMap(themeDisplay.getSiteDefaultLocale(), name),
-			getLocalizedMap(themeDisplay.getSiteDefaultLocale(), description),
-			ddmForm, ddmFormLayout, serviceContext);
-	}
-
-	protected DDLRecordSet updateRecordSet(
-			ActionRequest actionRequest, long ddmStructureId)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		long recordSetId = ParamUtil.getLong(actionRequest, "recordSetId");
-
-		String name = ParamUtil.getString(actionRequest, "name");
-		String description = ParamUtil.getString(actionRequest, "description");
-
-		ServiceContext serviceContext = ServiceContextFactory.getInstance(
-			DDLRecordSet.class.getName(), actionRequest);
-
-		return ddlRecordSetService.updateRecordSet(
-			recordSetId, ddmStructureId,
-			getLocalizedMap(themeDisplay.getSiteDefaultLocale(), name),
-			getLocalizedMap(themeDisplay.getSiteDefaultLocale(), description),
-			DDLRecordSetConstants.MIN_DISPLAY_ROWS_DEFAULT, serviceContext);
-	}
+	@Reference
+	protected AddUpdateRecordSetMVCCommandHelper
+		addUpdateRecordSetMVCCommandHelper;
 
 }
